build signup function

+++ src/main/java/com/example/web/domain/user/UserRepository.java
@@ -2,6 +2,5 @@ package com.example.web.domain.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<SiteUser, Long> {
-
+public interface UserRepository extends JpaRepository<SiteUser, Integer> {
 }
